requirements.txt
absl-py==2.1.0
accelerate==1.3.0
albumentations==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1652035089999/work
efficientnet-pytorch==0.7.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
face-recognition @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_1619808233029/work
face-recognition-models @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_models_1595707436534/work
fastapi==0.108.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2023.12.2
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.28.1
idna==3.6
imageio==2.33.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1717717528849/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1719582526268/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.5
joblib==1.3.2
jpegio==0.2.4
jsonpickle==3.0.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257277185/work
keras==3.1.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
munch==4.0.0
namex==0.0.7
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
optree==0.15.0
packaging==23.2
pandas==2.2.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pdf2image==1.17.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pretrainedmodels==0.7.4
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==4.23.4
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pyDeprecate==0.3.2
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.6
python-whois==0.8.0
pytz==2024.1
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq @ file:///croot/pyzmq_1705605076900/work
qrdet==2.4
qreader==3.12
quadrilateral-fitter==1.8
qudida==0.0.4
regex==2024.11.6
requests==2.31.0
rich==13.7.1
safetensors==0.5.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
segmentation-models-pytorch==0.2.1
shapely==2.0.3
six==1.15.0
smmap==5.0.1
sniffio==1.3.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.32.0.post1
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
tokenizers==0.21.0
torch==2.7.0
torchvision==0.22.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722796999/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.48.3
triton==3.3.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
ultralytics==8.1.19
urllib3==2.1.0
uvicorn==0.25.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
wrapt==1.16.0
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
